restructure utils to centralize shareable functions
