+++ web/lib/util.ts
@@ -1,18 +1,6 @@
 import randomColor from 'randomcolor';
-import { randomBytes } from 'crypto';
 import { GRAB_PATH } from 'config/paths';
 import { UAParser } from 'ua-parser-js';
-
-export const bufferFromString = (input: string) => {
-    const size = input.length;
-
-    const buffer = new ArrayBuffer(size);
-    const viewArray = new Uint8Array(buffer);
-
-    for (let i = 0; i < size; i++) viewArray[i] = input.charCodeAt(i);
-
-    return viewArray;
-};
 
 // ref: https://github.com/davidmerfield/randomColor#options
 const hues = [
@@ -49,9 +37,5 @@ export const generateGrabUrl = (id: string) => {
 };
 
 export const generateDropKey = (id: string) => `drop:${id}`;
-
-export const generateIV = () => randomBytes(12).toString('binary');
-
-export const getIVBuffer = (iv: string) => Buffer.from(iv, 'binary');
 
 export const getBrowserInfo = () => new UAParser().getResult();
